Moves custom files handler
